fix: validate session with remote
Before, we did not check whether the local session was still valid with the
remote, which lead to 401 errors.

Now validate the session and only grant offline login if we know so.

Note, we may still see 401 errors when the user comes back online if the session
has expired in the meantime. Consumers should call `canActivate` after verifying
the user has come back online and act accordingly.

BREAKING CHANGE: requires getSession to be exposed on sessionService. See
[pouchdb-authentication#dbgetSession](https://github.com/nolanlawson/pouchdb-authentication#dbgetsessionopts--callback)
for an example implementation.

Closes #11.